internal: Move helpers into Kubernetes runtime
The materialization helpers were shared by the local and Kubernetes adapters.
After removing the local adapter, Kubernetes is their only consumer and owns
all of the behavior they implement.

Move the helper implementation and tests into the concrete runtime package so
the directory layout matches that ownership and no obsolete shared utility
package remains.
