+++ internal/registry/runtimes/kubernetes/materialization_test.go
@@ -1,4 +1,4 @@
-package utils
+package kubernetes
 
 import (
 	"context"
